+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2026-05-03
+
+### Added
+
+- Add a dialog for creating or editing an item.
+
 ## [0.1.0] - 2026-03-21
 
 ### Added